+++ src/main/java/com/mandarina/game/ui/VolumeButtonCts.java
@@ -1,7 +1,7 @@
 package com.mandarina.game.ui;
 
 public class VolumeButtonCts {
-	public static final double VOLUME_VALUE_DEFAULT = 0.2;
+	public static final float VOLUME_VALUE_DEFAULT = 0.2f;
 	public static final int VOLUME_WIDTH_DEFAULT = 28;
 	public static final int VOLUME_HEIGHT_DEFAULT = 44;
 	public static final int SLIDER_WIDTH_DEFAULT = 215;
